I have added the functions for the forecast display
this time I have used the API to call the forecast 6 times on intervals of 3 hours and then created a function to show each block of 3 hours with the time, the appropiate icon and the min and max temperature in celsius. I have also formatted the hour display. Also I changed my original function to display the date at the top to make it more customised then just using the toDateString method.